@@ -1,53 +1,37 @@
 {
   "name": "atkjs-ui",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "description": "Agile Toolkit Javascript library.",
   "main": "lib/atkjs-ui.js",
   "scripts": {
-    "build": "webpack --env build",
-    "dev": "webpack --progress --colors --watch --env dev",
-    "test": "mocha --compilers js:babel-core/register --colors ./test/*.spec.js",
-    "test:watch": "mocha --compilers js:babel-core/register --colors -w ./test/*.spec.js"
-  },
-  "devDependencies": {
-    "@types/jquery": "^2.0.41",
-    "babel": "6.5.2",
-    "babel-core": "6.22.1",
-    "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.10",
-    "babel-plugin-add-module-exports": "^0.3.2",
-    "babel-preset-es2015": "6.22.0",
-    "chai": "3.4.1",
-    "eslint": "^5.2.0",
-    "eslint-loader": "1.6.1",
-    "mocha": "^5.2.0",
-    "webpack": "2.2.1",
-    "yargs": "6.6.0"
+    "build": "webpack --env.production",
+    "dev": "webpack --progress --colors --watch --env.development",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/atk4/ui"
+    "url": "https://github.com/atk4/ui/tree/develop/js"
   },
-  "keywords": [
-    "webpack",
-    "es6",
-    "starter",
-    "library",
-    "universal",
-    "umd",
-    "commonjs"
-  ],
   "author": "Agile Toolkit",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/atk4/ui/issues"
   },
   "homepage": "http://www.agiletoolkit.org/",
+  "devDependencies": {
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.1.6",
+    "@babel/preset-env": "^7.1.6",
+    "babel-loader": "^8.0.4",
+    "eslint": "^5.9.0",
+    "uglifyjs-webpack-plugin": "^2.0.1",
+    "webpack": "^4.26.1",
+    "webpack-cli": "^3.1.2"
+  },
   "dependencies": {
+    "@babel/polyfill": "^7.0.0",
     "column-resizer": "^1.1.2",
     "debounce": "^1.1.0",
-    "jquery": "^3.1.1",
-    "locutus": "^2.0.9",
-    "semantic-ui-modal": "^2.2.10"
+    "locutus": "^2.0.9"
   }
 }
